+++ src/numba_cuda_mlir/numba_cuda/np/arrayobj.py
@@ -4986,102 +4986,6 @@ def generate_getitem_setitem_with_axis(ndim, kind):
     return register_jitable(fn)
 
 
-@overload(np.take)
-@overload_method(types.Array, "take")
-def numpy_take(a, indices, axis=None):
-    if cgutils.is_nonelike(axis):
-        if isinstance(a, types.Array) and isinstance(indices, types.Integer):
-
-            def take_impl(a, indices, axis=None):
-                if indices > (a.size - 1) or indices < -a.size:
-                    raise IndexError("Index out of bounds")
-                return a.ravel()[indices]
-
-            return take_impl
-
-        if isinstance(a, types.Array) and isinstance(indices, types.Array):
-            F_order = indices.layout == "F"
-
-            def take_impl(a, indices, axis=None):
-                ret = np.empty(indices.size, dtype=a.dtype)
-                if F_order:
-                    walker = indices.copy()  # get C order
-                else:
-                    walker = indices
-                it = np.nditer(walker)
-                i = 0
-                flat = a.ravel()
-                for x in it:
-                    if x > (a.size - 1) or x < -a.size:
-                        raise IndexError("Index out of bounds")
-                    ret[i] = flat[x]
-                    i = i + 1
-                return ret.reshape(indices.shape)
-
-            return take_impl
-
-        if isinstance(a, types.Array) and isinstance(indices, (types.List, types.BaseTuple)):
-
-            def take_impl(a, indices, axis=None):
-                convert = np.array(indices)
-                return np.take(a, convert)
-
-            return take_impl
-    else:
-        if isinstance(a, types.Array) and isinstance(indices, types.Integer):
-            t = (0,) * (a.ndim - 1)
-
-            # np.squeeze is too hard to implement in Numba as the tuple "t"
-            # needs to be allocated beforehand we don't know it's size until
-            # code gets executed.
-            @register_jitable
-            def _squeeze(r, axis):
-                tup = tuple(t)
-                j = 0
-                assert axis < len(r.shape) and r.shape[axis] == 1, r.shape
-                for idx in range(len(r.shape)):
-                    s = r.shape[idx]
-                    if idx != axis:
-                        tup = tuple_setitem(tup, j, s)
-                        j += 1
-                return r.reshape(tup)
-
-            def take_impl(a, indices, axis=None):
-                r = np.take(a, (indices,), axis=axis)
-                if a.ndim == 1:
-                    return r[0]
-                if axis < 0:
-                    axis += a.ndim
-                return _squeeze(r, axis)
-
-            return take_impl
-
-        if isinstance(a, types.Array) and isinstance(
-            indices, (types.Array, types.List, types.BaseTuple)
-        ):
-            ndim = a.ndim
-
-            _getitem = generate_getitem_setitem_with_axis(ndim, "getitem")
-            _setitem = generate_getitem_setitem_with_axis(ndim, "setitem")
-
-            def take_impl(a, indices, axis=None):
-                if axis < 0:
-                    axis += a.ndim
-
-                if axis < 0 or axis >= a.ndim:
-                    msg = f"axis {axis} is out of bounds for array of dimension {a.ndim}"
-                    raise ValueError(msg)
-
-                shape = tuple_setitem(a.shape, axis, len(indices))
-                out = np.empty(shape, dtype=a.dtype)
-                for i in range(len(indices)):
-                    y = _getitem(a, indices[i], axis)
-                    _setitem(out, i, axis, y)
-                return out
-
-            return take_impl
-
-
 def _arange_dtype(*args):
     bounds = [a for a in args if not isinstance(a, types.NoneType)]
